pyproject.toml
[project]
name = "crcon-qqbot"
version = "1.0.0"
description = "QQ机器人对接人间地狱CRCON管理面板API"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "nonebot2>=2.2.1",
    "nonebot-adapter-onebot>=2.4.4",
    "nonebot-plugin-apscheduler>=0.4.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.1",
    "pydantic>=2.5.2",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true